refactor: переименовывает файлы в соответствии с критериями
